refactor(utils/getdirname): move `getDirName()` to `src/`
