+++ tests/packageParser.js
@@ -2,7 +2,7 @@ import test from 'ava';
 import { resolve as resolvePath } from 'pathe';
 import mockFS from 'mock-fs';
 import packageParser from '../src/packageParser.js';
-import getDirName from './__helpers__/getDirName.js';
+import getDirName from '../src/utils/getDirName.js';
 
 const __dirname = getDirName( import.meta.url );
 const fixtures = {
